+++ src/main/java/com/dnd/MusicLog/global/jwt/dto/AuthTokensResponseDto.java
@@ -1,4 +1,4 @@
-package com.dnd.MusicLog.user.dto;
+package com.dnd.MusicLog.global.jwt.dto;
 
 import lombok.*;
